+++ app/collections/[uuid]/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Metadata } from "next";
 import PageLayout from "../../src/components/pageLayout/pageLayout";
-import { createAdobeAnalyticsPageName } from "@/src/utils/utils";
+import { createAdobeAnalyticsPageName, imageURL } from "@/src/utils/utils";
 import CollectionPage from "@/src/components/pages/collectionPage/collectionPage";
 import { CollectionsApi } from "@/src/utils/apiClients";
 import CollectionSearchParamsType from "@/src/types/CollectionSearchParams";
@@ -15,10 +15,21 @@ export async function generateMetadata({
   params,
 }: CollectionProps): Promise<Metadata> {
   const slug = params.uuid;
+  const collectionData = await CollectionsApi.getCollectionData(slug);
+  const title = collectionData?.title ?? slug;
+
   return {
-    title: `${slug} - NYPL Digital Collections`,
+    title: `${title} - NYPL Digital Collections`,
     openGraph: {
-      title: `${slug} - NYPL Digital Collections`,
+      title: `${title} - NYPL Digital Collections`,
+      images: [
+        {
+          url: collectionData.imageID
+            ? imageURL(collectionData.imageID, "full", "!288,288", "0")
+            : "https://digitalcollections.nypl.org/featured_items/ps_mss_831.jpg",
+          alt: `${title} - NYPL Digital Collections`,
+        },
+      ],
     },
   };
 }
@@ -45,7 +56,7 @@ export default async function Collection({
 
   const searchResults = await CollectionsApi.getSearchData({
     keyword: searchParams.q,
-    sort: searchParams.sort,
+    sort: searchParams.sort ? searchParams.sort : "sequence",
     page: searchParams.page,
     filters,
   });