+++ app/src/models/collection.ts
@@ -3,8 +3,8 @@ export class CollectionModel {
   uuid: string;
   abstract?: string;
   accessCondition?: string[];
-  archivesCollectionID?: string;
-  bNumber?: string;
+  mssID?: string;
+  bNumbers?: string[];
   contentNote?: string[];
   dateCaptured: string;
   dateCreated: string;
@@ -37,8 +37,8 @@ export class CollectionModel {
     this.uuid = data.uuid;
     this.abstract = data.abstract;
     this.accessCondition = data.accessCondition;
-    this.archivesCollectionID = data.archivesCollectionID;
-    this.bNumber = data.bNumber;
+    this.mssID = data.mssID;
+    this.bNumbers = data.bnumber?.split(" ");
     this.contentNote = data.contentNote;
     this.dateCaptured = data.dateCaptured;
     this.dateCreated = data.dateCreated;